Added a new format code %proxy_protocol::>h, pt 2
Made all PROXY v2 TLV log formatting similar to %http::>h, allowing
filtering by TLV type or TLV value element.

Also:

* documented ProxyProtocol methods

* eliminated code duplication in HttpHeader::getByNameListMember()
  and HttpHeader::getListMember(), moving the common part into a
  separate getListMember()

+++ src/format/Token.h
@@ -50,7 +50,7 @@ public:
     const char *label;
     struct {
         char *string;
-        unsigned int proxyProtocolType = 0;
+        uint8_t proxyProtocolType = 0;
 
         struct {
             char *header;
